+++ src/java/DataAccesObject/Empleado_DAO.java
@@ -684,7 +684,7 @@ public class Empleado_DAO {
                             Det_Orden_DTO det = new Det_Orden_DTO();
                             det.setId_det_orden(rs.getInt("id_Det_Orden"));
                             det.setEstudio(E.getEst_Uni(rs.getInt("id_Est_Uni")));
-                            det.setDescuento(rs.getInt("Descuento"));
+                            det.setDescuento(rs.getFloat("Descuento"));
                             det.setFecha_Entrega(rs.getString("Fecha_Entrega"));
                             det.setT_Entrega(rs.getString("Tipo_Entrega"));
                             det.setSubtotal(rs.getFloat("Subtotal"));
